Added ModelState check in OnPost in Delete.cshtml.cs

+++ src/Pages/Product/Delete.cshtml.cs
@@ -56,6 +56,11 @@ namespace ContosoCrafts.WebSite.Pages.Product
         /// <returns></returns>
         public IActionResult OnPost()
         {
+            if (!ModelState.IsValid)
+            {
+                return Page();
+            }
+
             ProductService.DeleteData(Product.Id);
 
             return RedirectToPage("./Index");
